% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_BIOMATE_bibstyle.R
\name{add_bibstyle}
\alias{add_bibstyle}
\title{Create a new bibstyle for use in BIOMATE.}
\usage{
add_bibstyle()
}
\value{
returns message letting the user know the task was completed.
}
\description{
Executing this function will add a new "BIOMATE" style to a local library. This style is called within PIG_to_WHPE() and PROF_to_WHPE().
}
\author{
Kimberlee Baldry
}
