% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export_ref.R
\name{export_ref}
\alias{export_ref}
\title{Reference BIO-MATE right}
\usage{
export_ref(EXPOCODES = "all", streams = "all", reports = "Y", path_out = "./")
}
\description{
Export the BIO-MATE citation, along with individual data citations in on BibTeX file to make referencing easy. Upload this file into your reference manager or text editing software. A handy reference table is also compiles. Remember all individual data sets that are downloaded with BIO-MATE need to be cited appropriately.

 @return A BibTeX file, saved to the directed location, and a data.table containing reference codes for datasets.

 @param EXPOCODE A list of the EXPOCODES for the data you have used. Deefault is "all".
 @param streams The data streams to be cited. Can be "all" or a combination of "PIG", "PROF","POC' and "UWY".
 @param reports Are the cruise report citations needed: Y/N.
 @param path_out The path where the BibTeX file will be saved.
}
\author{
Kimberlee Baldry
}
