% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/POC_to_WHPE.R
\name{POC_to_WHPE}
\alias{POC_to_WHPE}
\title{Reformat particulate organic carbon (poc) data to WHP-Exchange format with MAREDAT headers}
\usage{
POC_to_WHPE(
  file_path,
  path_out,
  userID = "IMASUTASKB",
  row_start = 1,
  row_end = NA,
  t_thresh = 6,
  d_thresh = 1000,
  version = "1.0"
)
}
\arguments{
\item{file_path}{The path to the processing metadata file, titled "POC_meta.csv".}

\item{path_out}{The path where reformatted files are to be sent. A subdirectory "poc" will be created in here, if one does not already exist.}

\item{userID}{An ID code for the user creating the files.}

\item{row_start}{The start row of the processing metadata table, for batch run worflows.}

\item{row_end}{The end row of the processing metadata table, for batch run worflows.}
}
\value{
returns a successful run message and saves reformatted data to a "poc" subdirectory in path_out.
}
\description{
BIOMATE reformats particulate organic carbon (poc) files and saves them as a file that is tagged by EXPOCODE, METHOD and SOURCE. The resulting file follows WHP-Exchange format guidelines with the exception that MAREDAT pigment headers are used. Information is required from the user, stored within a user-entered metadata file particularly for batch runs, or provided via the BIOMATE submission app. The metadata file must be named "POC_meta.csv". All poc files must be split by EXPOCODE (can be multiple files in the same folder with identical extensions). Use the split_delim_file function to convert files if there are multiple voyages in a single file.
}
\note{
This function only supports well structured text delimited and NetCDF file formats. It is almost identical yo PIG_to_WHPE, despite alterations to pigment/poc header names, file paths and error messages.
}
\author{
Kimberlee Baldry
}
