% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PROF_to_WHPE.R
\name{PROF_to_WHPE}
\alias{PROF_to_WHPE}
\title{Reformat profiling sensor (CTD) data to WHP-Exchange format.}
\usage{
PROF_to_WHPE(
  file_path,
  path_out,
  userID = "IMASUTASKB",
  row_start = 1,
  row_end = NA,
  trace_file = F,
  version = "1.0"
)
}
\arguments{
\item{file_path}{The path to the processing metadata file, titled "PROF_meta.csv".}

\item{path_out}{The path where reformatted files are to be sent. A subdirectory "profiling_sensors" will be created in here, if one does not already exist.}

\item{userID}{An ID code for the user creating the files.}

\item{row_start}{The start row of the processing metadata table, for batch run worflows.}

\item{row_end}{The end row of the processing metadata table, for batch run worflows.}

\item{trace_file}{TRUE will print out each file name as it is reformatted, FALSE suppresses this. It helps with de-bugging and identifying formatting errors during development.}
}
\value{
returns a successful run message and saves reformatted data to a "profiling_sensors" subdirectory in path_out.
}
\description{
BIOMATE reformats profiling sensor files and saves them as a file that is tagged by EXPOCODE, STATION and CAST. The resulting file follows WHP-Exchange format guidelines. Information is required from the user, stored within a user-entered metadata file particularly for batch runs, or provided via the BIOMATE submission app. The metadata file must be named "PROF_meta.csv". All PROF/CTD files must be singular per cast with the same extention, in one folder. Use the split_delim_file function to convert files to this format.
}
\note{
This function only supports well structured text delimited and NetCDF file formats.
}
\author{
Kimberlee Baldry
}
