% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/split_compiled_delim_file.R
\name{split_delim_file}
\alias{split_delim_file}
\title{Split a delimited text file into sub-files based on EXPOCODES}
\usage{
split_delim_file(
  path,
  file_name,
  delim,
  line_start = 1,
  expo_split = T,
  synonym_var_name,
  station_split = F,
  station_var_name = NA,
  fillcell = F
)
}
\arguments{
\item{path}{The path where the file is located}

\item{file_name}{The name of the file to be split}

\item{delim}{The text delimiter of the file to be split}

\item{line_start}{The line in which the data table starts (i.e. the line the variable headers are on)}

\item{expo_split}{Set to T if working with EXPOCODE synonyms data to split further into EXPOCODES. Note you need a supplementry netadata file in this case, see description for more details.}

\item{synonym_var_name}{Only needed when expo_split = T. The variable, containing EXPOCODE synonyms, that the file is to be split by.}

\item{station_split}{Set to T if working with PROF/CTD data to split further into profileing stations. This will output multiple files to the output folder}

\item{station_var_name}{Only needed is station_split = T. The variable containing station IDs}

\item{fillcell}{logical. If TRUE then in case the rows have unequal length, blank fields are implicitly added. See ‘fill’ and "Details" in "read.table" for more information.}
}
\value{
returns a "successful run" message and saves split files in a subdirectory named "split" in the location of the original data file.
}
\description{
BIOMATE requires a minimum of one file per EXPOCODE for the PIG, UWY and POC data streams and one file per STATION for the PROF data stream. Sometimes data publishers deliver text files differently, lumping information from multiple EXPOCODES or stations together. In order for BIOMATE to handle data delivered in this way, it first needs to be split appropriately. If data is to be split by EXPOCODES (expo_split = T), an acompaning EXPOCODE synonym look up table is required to inform the software. This file must be named "[file_name]_EXPOCODE_synonym.csv" where file_name is the name of the file to be split it must be a text, comma delimited file with two columns. The first column named "EXPOCODE" should contain the EXPOCODE as used in the data aggregation. The second column named "synonym" muct contain the EXPOCODE synonym that appears in the file. An example file can be found here (link). A new folder will be created by the script in the loction of the file (path) for each EXPOCODE. The split files will be saved as comma delimited text files in their respective locations.
}
\author{
Kimberlee Baldry
}
