% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Marlin_SS_split_CTD.R
\name{SS_CTDsplit}
\alias{SS_CTDsplit}
\title{Split old Southern Surveyor files}
\usage{
SS_CTDsplit(path)
}
\arguments{
\item{path}{The path where the Marlin data is stored. This path must include sub-folders for expocodes}
}
\value{
returns a "successful run" message and saves split files in the same directory as the original data file.
}
\description{
This script splits old southern surveyor text files from ctds, and stors tab separated files (.tsv) in the same directory as the original file. It references a "Marlin_SS_split_CTD_meta.csv" file. This file must be stored in the path directory. It will output a data amendment file for records.
}
\note{
This function is only useful to reformat very old Southern Surveyor files.
}
\author{
Kimberlee Baldry
}
