% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UWY_to_WHPE.R
\name{UWY_to_WHPE}
\alias{UWY_to_WHPE}
\title{Reformat underway sensor data to WHP-Exchange format}
\usage{
UWY_to_WHPE(
  file_path,
  path_out,
  userID = "IMASUTASKB",
  row_start = 1,
  row_end = NA,
  t_thresh = 6,
  d_thresh = 1000,
  version = "1.0"
)
}
\arguments{
\item{file_path}{The path to the processing metadata file, titled "UWY_meta.csv".}

\item{path_out}{The path where reformatted files are to be sent. A subdirectory "underway_sensors" will be created in here, if one does not already exist.}

\item{userID}{An ID code for the user creating the files.}

\item{row_start}{The start row of the processing metadata table, for batch run worflows.}

\item{row_end}{The end row of the processing metadata table, for batch run worflows.}
}
\value{
returns a successful run message and saves reformatted data to a "underway_sensors" subdirectory in path_out.
}
\description{
BIOMATE reformats underway sensor files and saves them as a file that is tagged by EXPOCODE. The resulting file follows WHP-Exchange format guidelines. Information is required from the user, stored within a user-entered metadata file particularly for batch runs, or provided via the BIOMATE submission app. The metadata file must be named "UWY_meta.csv". Multiple, sequential underway sensor files can exist for each EXPOCODE, but one file cannot contain data from multiple EXPOCODES. Use the split_delim_file function to convert files if there are multiple voyages in a single file.
}
\note{
This function only supports well structured text delimited and NetCDF file formats.
}
\author{
Kimberlee Baldry
}
